Finish moving implementation over to .cpp
